+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 -   For pyspark projects, allows to configure post script for dataproc initialization
+-   FIX: Broken support for kedro<0.17
 
 ## [0.7.1] - 2021-10-21